pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hfqt"
version = "0.1.0"
description = "Minimal event-driven AI trading skeleton with local paper and Futu simulated broker support."
readme = "docs/5090_local_quant_trading_plan.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110,<1",
  "futu-api>=10,<11",
  "httpx>=0.27,<1",
  "pydantic>=2.10,<3",
  "pycryptodome>=3.20,<4",
  "uvicorn>=0.30,<1",
  "yfinance>=0.2.61,<1",
]

[project.scripts]
hfqt = "hfqt.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
